refactor(py): make our code installable as a package
This makes it possible to install our Python code as a package with `pip`,
including an automatically generated helper script called
`pathfinder_python_worker` that can be used to start up our subprocess.

This way we can get rid of including `call.py` into our Rust executable
and still have a reliable way of executing the subprocess.
