+++ py/.gitignore
@@ -1,2 +1,4 @@
+build/
 .venv/
 **/__pycache__
+*.egg-info/
