+++ py/pyproject.toml
@@ -1,0 +1,26 @@
+[build-system]
+requires = ["setuptools>=65.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pathfinder-worker"
+version = "0.1.0"
+description = "Pathfinder Python worker subprocess"
+readme = "README.md"
+dependencies = [
+    "cairo-lang == 0.10.1",
+    "zstandard == 0.18.0",
+    "cachetools == 5.0.0",
+]
+requires-python = ">=3.8"
+
+[project.optional-dependencies]
+dev = [
+    "pip-tools",
+    "pytest == 7.1.3",
+    "flake8 == 5.0.4",
+    "black == 22.10.0",
+]
+
+[project.scripts]
+pathfinder_python_worker = "pathfinder_worker.call:main"
