+++ py/requirements-dev.txt
@@ -1,78 +1,238 @@
 #
-# This file is autogenerated by pip-compile with python 3.9
+# This file is autogenerated by pip-compile with python 3.10
 # To update, run:
 #
-#    pip-compile requirements-dev.in
+#    pip-compile --extra=dev --output-file=requirements-dev.txt pyproject.toml
 #
+aiohttp==3.8.3
+    # via
+    #   cairo-lang
+    #   web3
+aiosignal==1.2.0
+    # via aiohttp
+async-timeout==4.0.2
+    # via aiohttp
 attrs==22.1.0
     # via
-    #   -c requirements.txt
+    #   aiohttp
+    #   jsonschema
     #   pytest
+base58==2.1.1
+    # via multiaddr
+bitarray==2.6.0
+    # via eth-account
 black==22.10.0
-    # via -r requirements-dev.in
+    # via pathfinder-worker (pyproject.toml)
 build==0.8.0
     # via pip-tools
+cachetools==5.0.0
+    # via
+    #   cairo-lang
+    #   pathfinder-worker (pyproject.toml)
+cairo-lang==0.10.1
+    # via pathfinder-worker (pyproject.toml)
+certifi==2022.9.24
+    # via requests
+charset-normalizer==2.1.1
+    # via
+    #   aiohttp
+    #   requests
 click==8.1.3
     # via
     #   black
     #   pip-tools
-flake8==5.0.4
-    # via -r requirements-dev.in
-iniconfig==1.1.1
+cytoolz==0.12.0
     # via
-    #   -c requirements.txt
-    #   pytest
+    #   eth-keyfile
+    #   eth-utils
+ecdsa==0.18.0
+    # via cairo-lang
+eth-abi==2.2.0
+    # via
+    #   eth-account
+    #   web3
+eth-account==0.5.9
+    # via web3
+eth-hash[pycryptodome]==0.3.3
+    # via
+    #   cairo-lang
+    #   eth-utils
+    #   web3
+eth-keyfile==0.5.1
+    # via eth-account
+eth-keys==0.3.4
+    # via
+    #   eth-account
+    #   eth-keyfile
+eth-rlp==0.2.1
+    # via
+    #   eth-account
+    #   web3
+eth-typing==2.3.0
+    # via
+    #   eth-abi
+    #   eth-keys
+    #   eth-utils
+    #   web3
+eth-utils==1.10.0
+    # via
+    #   eth-abi
+    #   eth-account
+    #   eth-keyfile
+    #   eth-keys
+    #   eth-rlp
+    #   rlp
+    #   web3
+fastecdsa==2.2.3
+    # via cairo-lang
+flake8==5.0.4
+    # via pathfinder-worker (pyproject.toml)
+frozendict==1.2
+    # via cairo-lang
+frozenlist==1.3.1
+    # via
+    #   aiohttp
+    #   aiosignal
+hexbytes==0.3.0
+    # via
+    #   eth-account
+    #   eth-rlp
+    #   web3
+idna==3.4
+    # via
+    #   requests
+    #   yarl
+iniconfig==1.1.1
+    # via pytest
+ipfshttpclient==0.8.0a2
+    # via web3
+jsonschema==4.16.0
+    # via web3
+lark==1.1.3
+    # via cairo-lang
+lru-dict==1.1.8
+    # via web3
+marshmallow==3.18.0
+    # via
+    #   cairo-lang
+    #   marshmallow-dataclass
+    #   marshmallow-enum
+    #   marshmallow-oneofschema
+marshmallow-dataclass==8.5.9
+    # via cairo-lang
+marshmallow-enum==1.5.1
+    # via cairo-lang
+marshmallow-oneofschema==3.0.1
+    # via cairo-lang
 mccabe==0.7.0
     # via flake8
+mpmath==1.2.1
+    # via
+    #   cairo-lang
+    #   sympy
+multiaddr==0.0.9
+    # via ipfshttpclient
+multidict==6.0.2
+    # via
+    #   aiohttp
+    #   yarl
 mypy-extensions==0.4.3
     # via
-    #   -c requirements.txt
     #   black
+    #   typing-inspect
+netaddr==0.8.0
+    # via multiaddr
+numpy==1.23.4
+    # via cairo-lang
 packaging==21.3
     # via
-    #   -c requirements.txt
     #   build
+    #   marshmallow
     #   pytest
+parsimonious==0.8.1
+    # via eth-abi
 pathspec==0.10.1
     # via black
 pep517==0.13.0
     # via build
 pip-tools==6.9.0
-    # via -r requirements-dev.in
+    # via pathfinder-worker (pyproject.toml)
+pipdeptree==2.3.1
+    # via cairo-lang
 platformdirs==2.5.2
     # via black
 pluggy==1.0.0
-    # via
-    #   -c requirements.txt
-    #   pytest
+    # via pytest
+prometheus-client==0.15.0
+    # via cairo-lang
+protobuf==3.19.5
+    # via web3
 py==1.11.0
-    # via
-    #   -c requirements.txt
-    #   pytest
+    # via pytest
 pycodestyle==2.9.1
     # via flake8
+pycryptodome==3.15.0
+    # via
+    #   eth-hash
+    #   eth-keyfile
 pyflakes==2.5.0
     # via flake8
 pyparsing==3.0.9
-    # via
-    #   -c requirements.txt
-    #   packaging
+    # via packaging
+pyrsistent==0.18.1
+    # via jsonschema
 pytest==7.1.3
     # via
-    #   -c requirements.txt
-    #   -r requirements-dev.in
+    #   cairo-lang
+    #   pathfinder-worker (pyproject.toml)
+    #   pytest-asyncio
+pytest-asyncio==0.20.0
+    # via cairo-lang
+pyyaml==6.0
+    # via cairo-lang
+requests==2.28.1
+    # via
+    #   ipfshttpclient
+    #   web3
+rlp==2.0.1
+    # via
+    #   eth-account
+    #   eth-rlp
+six==1.16.0
+    # via
+    #   ecdsa
+    #   multiaddr
+    #   parsimonious
+sympy==1.11.1
+    # via cairo-lang
 tomli==2.0.1
     # via
-    #   -c requirements.txt
     #   black
     #   build
+    #   pep517
     #   pytest
+toolz==0.12.0
+    # via cytoolz
+typeguard==2.13.3
+    # via cairo-lang
 typing-extensions==4.4.0
-    # via
-    #   -c requirements.txt
-    #   black
+    # via typing-inspect
+typing-inspect==0.8.0
+    # via marshmallow-dataclass
+urllib3==1.26.12
+    # via requests
+varint==1.0.2
+    # via multiaddr
+web3==5.31.1
+    # via cairo-lang
+websockets==9.1
+    # via web3
 wheel==0.37.1
     # via pip-tools
+yarl==1.8.1
+    # via aiohttp
+zstandard==0.18.0
+    # via pathfinder-worker (pyproject.toml)
 
 # The following packages are considered to be unsafe in a requirements file:
 # pip
