+++ py/tools/compute_block_hash.py
