+++ py/tools/compute_class_hash.py
